Automation/requirements.txt
lxml
pandas
paramiko
ping3
pysqlite3
pytest
python-jenkins
requests
selenium
transitions
wheel
